Revert "frontend modularization, update gitignore"
