+++ .gitignore
@@ -7,7 +7,7 @@ node_modules/
 coverage/
 
 # Next.js
-.next/
+/.next/
 /out/
 
 # Production
@@ -39,5 +39,5 @@ yarn-error.log*
 next-env.d.ts
 
 # Backend specific
-backend/bin/
+backend/bin/www.js
 *.log
