--- frontend/app/components/Error.tsx
@@ -1,11 +1,0 @@
-import { ErrorBox } from "../styles/components/Error.styles";
-
-const Error = ({ error, darkMode }) => {
-  return (
-    <>
-      <ErrorBox $darkMode={darkMode}>{error}</ErrorBox>
-    </>
-  );
-};
-
-export default Error;
