--- frontend/app/components/Loader.tsx
@@ -1,12 +1,0 @@
-import { LoaderWrapper, Spinner } from "../styles/components/Loader.styles";
-const Loader = () => {
-  return (
-    <>
-      <LoaderWrapper>
-        <Spinner />
-      </LoaderWrapper>
-    </>
-  );
-};
-
-export default Loader;
