+++ frontend/app/layout.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import "./styles/globals.css";
 
 export default function RootLayout({
   children,
