--- frontend/app/styles/globals.css
@@ -1,3 +1,0 @@
-body {
-  margin: 0;
-}
