--- frontend/app/utils/getDayName.ts
@@ -1,6 +1,0 @@
-const getDayName = (dateStr: string) => {
-  const date = new Date(dateStr);
-  return date.toLocaleDateString("en-US", { weekday: "short" });
-};
-
-export default getDayName;
